agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "CopilotKit + Pydantic AI with Neon database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn",
    "pydantic-ai-slim[ag-ui]",
    "pydantic-ai-slim[google]",
    "python-dotenv",
    "logfire>=4.10.0",
    "psycopg2-binary",
]
